add inetaddress and fix ipv6
