+++ httpserver/socket.hpp
@@ -1,6 +1,7 @@
 #ifndef HTTP_SREVER_SOCKET_HPP_
 #define HTTP_SREVER_SOCKET_HPP_
 
+#include "inetaddress.hpp"
 #include <sys/types.h>
 #include <sys/socket.h>
 #include <netinet/in.h>
@@ -20,47 +21,15 @@ public:
 
   Socket() : fd_(-1) {}
   Socket(int fd) : fd_(fd) {}
-  explicit Socket(const std::string& addr, const std::string& port) {
-    int ret;
-    struct addrinfo hints;
-    struct addrinfo *result;
+  Socket(const std::string& addr, const std::string& port) {
     struct addrinfo *rp;
-    memset(&hints, 0, sizeof(struct addrinfo));
-    bool ipv6 = true; 
-    if (ipv6) {
-      hints.ai_family = AF_INET6;
-      hints.ai_socktype = SOCK_STREAM;
-      hints.ai_protocol = 0;
-      hints.ai_canonname = NULL;
-      hints.ai_addr = NULL;
-      hints.ai_next = NULL;
-      ret = getaddrinfo(addr.c_str(), port.c_str(),&hints, &result);
-      if (ret != 0) {
-        fprintf(stderr, "getaddrinfo: %s\n", gai_strerror(ret));
-        exit(EXIT_FAILURE);
-      }
-    }
-    else {
-      hints.ai_family = AF_INET;
-      hints.ai_socktype = SOCK_STREAM;
-      hints.ai_protocol = 0;
-      hints.ai_canonname = NULL;
-      hints.ai_addr = NULL;
-      hints.ai_next = NULL;
-      ret = getaddrinfo(addr.c_str(), port.c_str(),&hints, &result);
-      if (ret != 0) {
-        fprintf(stderr, "getaddrinfo: %s\n", gai_strerror(ret));
-        exit(EXIT_FAILURE);
-      }
-    }
+    struct addrinfo* result = handleInetAddress(addr, port);
 
     for (rp = result; rp != NULL; rp = rp->ai_next) {
       fd_ = socket(rp->ai_family, rp->ai_socktype,
                    rp->ai_protocol);
       if (fd_ == -1)
         continue;
-      int reuse = 1;
-      ret = setsockopt( fd_, SOL_SOCKET, SO_REUSEADDR, &reuse, sizeof(reuse));
 
       if (bind(fd_ , rp->ai_addr, rp->ai_addrlen) == 0)
         break;                  /* Success */
